skills/microgrid-agent/reference/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "microgrid-agent"
version = "0.1.0"
description = "Open-source edge AI agent for autonomous microgrid management"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"

authors = [
    { name = "Broomva", email = "broomva@broomva.tech" },
]

keywords = [
    "microgrid",
    "renewable-energy",
    "edge-ai",
    "autonomous-agent",
    "knowledge-graph",
    "raspberry-pi",
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Operating System :: POSIX :: Linux",
]

dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
    "aiosqlite>=0.19",
    "aiohttp>=3.9",
    "paho-mqtt>=1.6,<2.0",
]

[project.optional-dependencies]
# Hardware protocols (RS-485 Modbus, Victron VE.Direct)
hardware = [
    "pymodbus>=3.5",
    "pyserial-asyncio>=0.6",
]

# TFLite runtime for ARM devices (Raspberry Pi)
arm = [
    "tflite-runtime",
]

# Development and testing
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.5",
]

# All optional dependencies
all = [
    "microgrid-agent[hardware,dev]",
]

[project.scripts]
microgrid-agent = "main:main"

[project.urls]
Homepage = "https://github.com/broomva/microgrid-agent"
Repository = "https://github.com/broomva/microgrid-agent"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests that take significant time",
    "hardware: marks tests that require physical devices",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
